Split utils and graph for cleaner feature guard
