+++ write-fonts/src/graph.rs
@@ -1,65 +1,29 @@
 //! A graph for resolving table offsets
 
-#[cfg(feature = "tables")]
 use font_types::Uint24;
 
-use crate::write::TableData;
-#[cfg(feature = "tables")]
-use crate::{table_type::TableType, tables::layout::LookupType};
+use crate::{
+    object::{ObjectId, ObjectStore},
+    offsets::OffsetLen,
+    table_type::TableType,
+    tables::layout::LookupType,
+    write::TableData,
+};
 
-#[cfg(feature = "tables")]
+use std::collections::HashMap;
 use std::collections::{BTreeMap, BTreeSet, BinaryHeap, HashSet, VecDeque};
-use std::{collections::HashMap, sync::atomic::AtomicU64};
 
 #[cfg(feature = "dot2")]
 mod graphviz;
-#[cfg(feature = "tables")]
 mod splitting;
 
-static OBJECT_COUNTER: AtomicU64 = AtomicU64::new(0);
-
-/// An identifier for an object in the compilation graph.
-#[derive(Debug, Clone, Copy, PartialOrd, Ord, Hash, PartialEq, Eq)]
-pub struct ObjectId(u64);
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-#[repr(u8)]
-pub enum OffsetLen {
-    Offset16 = 2,
-    Offset24 = 3,
-    Offset32 = 4,
-}
-
-impl OffsetLen {
-    /// The maximum value for an offset of this length.
-    #[cfg(feature = "tables")]
-    pub const fn max_value(self) -> u32 {
-        match self {
-            Self::Offset16 => u16::MAX as u32,
-            Self::Offset24 => (1 << 24) - 1,
-            Self::Offset32 => u32::MAX,
-        }
-    }
-}
-
-impl std::fmt::Display for OffsetLen {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match self {
-            Self::Offset16 => write!(f, "Offset16"),
-            Self::Offset24 => write!(f, "Offset24"),
-            Self::Offset32 => write!(f, "Offset32"),
-        }
-    }
-}
 /// A ranking used for sorting the graph.
 ///
 /// Nodes are assigned a space, and nodes in lower spaces are always
 /// packed before nodes in higher spaces.
 #[derive(Debug, Clone, Copy, PartialOrd, Ord, Hash, PartialEq, Eq)]
-#[cfg(feature = "tables")]
 pub struct Space(u32);
 
-#[cfg(feature = "tables")]
 impl Space {
     /// A generic space for nodes reachable via 16-bit offsets.
     const SHORT_REACHABLE: Space = Space(0);
@@ -73,29 +37,11 @@ impl Space {
     }
 }
 
-impl ObjectId {
-    pub fn next() -> Self {
-        ObjectId(OBJECT_COUNTER.fetch_add(1, std::sync::atomic::Ordering::Relaxed))
-    }
-}
-
-#[derive(Debug, Default)]
-pub(crate) struct ObjectStore {
-    pub(crate) objects: HashMap<TableData, ObjectId>,
-}
-
-impl ObjectStore {
-    pub(crate) fn add(&mut self, data: TableData) -> ObjectId {
-        *self.objects.entry(data).or_insert_with(ObjectId::next)
-    }
-}
-
 /// A graph of subtables, starting at a single root.
 ///
 /// This type is used during compilation, to determine the final write order
 /// for the various subtables.
 //NOTE: we don't derive Debug because it's way too verbose to be useful
-#[cfg(feature = "tables")]
 pub struct Graph {
     /// the actual data for each table
     objects: BTreeMap<ObjectId, TableData>,
@@ -111,7 +57,6 @@ pub struct Graph {
 }
 
 #[derive(Debug)]
-#[cfg(feature = "tables")]
 struct Node {
     size: u32,
     distance: u64,
@@ -124,7 +69,6 @@ struct Node {
 
 /// Scored used when computing shortest distance
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-#[cfg(feature = "tables")]
 struct Distance {
     // a space ranking; like rankings are packed together,
     // and larger rankings are packed after smaller ones.
@@ -136,12 +80,10 @@ struct Distance {
 
 //TODO: remove me? maybe? not really used right now...
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-#[cfg(feature = "tables")]
 struct Priority(u8);
 
 /// A record of an overflowing offset
 #[derive(Clone, Debug)]
-#[cfg(feature = "tables")]
 pub(crate) struct Overflow {
     parent: ObjectId,
     child: ObjectId,
@@ -149,7 +91,6 @@ pub(crate) struct Overflow {
     offset_type: OffsetLen,
 }
 
-#[cfg(feature = "tables")]
 impl Priority {
     const ZERO: Priority = Priority(0);
     const ONE: Priority = Priority(1);
@@ -164,7 +105,6 @@ impl Priority {
     }
 }
 
-#[cfg(feature = "tables")]
 impl Distance {
     const ROOT: Distance = Distance {
         space: Space::SHORT_REACHABLE,
@@ -177,7 +117,6 @@ impl Distance {
     }
 }
 
-#[cfg(feature = "tables")]
 impl Node {
     pub fn new(size: u32) -> Self {
         Node {
@@ -215,7 +154,6 @@ impl Node {
     }
 }
 
-#[cfg(feature = "tables")]
 impl Graph {
     pub(crate) fn from_obj_store(store: ObjectStore, root: ObjectId) -> Self {
         let objects = store.objects.into_iter().map(|(k, v)| (v, k)).collect();
@@ -315,7 +253,6 @@ impl Graph {
     /// returns `true` if a solution is found, `false` otherwise
     ///
     /// [repacker docs]: https://github.com/harfbuzz/harfbuzz/blob/main/docs/repacker.md
-    #[cfg(feature = "tables")]
     pub(crate) fn pack_objects(&mut self) -> bool {
         if self.basic_sort() {
             return true;
@@ -886,7 +823,6 @@ impl Graph {
         self.next_space
     }
 
-    #[cfg(feature = "tables")]
     fn try_promoting_subtables(&mut self) {
         let Some((can_promote, parent_id)) = self.get_promotable_subtables() else {
             return;
@@ -900,7 +836,6 @@ impl Graph {
         self.actually_promote_subtables(&to_promote);
     }
 
-    #[cfg(feature = "tables")]
     fn actually_promote_subtables(&mut self, to_promote: &[ObjectId]) {
         fn make_extension(type_: LookupType, subtable_id: ObjectId) -> TableData {
             const EXT_FORMAT: u16 = 1;
@@ -951,7 +886,6 @@ impl Graph {
     }
 
     // get the list of tables that can be promoted, as well as the id of their parent table
-    #[cfg(feature = "tables")]
     fn get_promotable_subtables(&self) -> Option<(Vec<ObjectId>, ObjectId)> {
         let can_promote = self
             .objects
@@ -1081,7 +1015,6 @@ impl Graph {
     /// Based on [`_presplit_subtables_if_needed`][presplit] in hb-repacker
     ///
     /// [presplit]: https://github.com/harfbuzz/harfbuzz/blob/5d543d64222c6ce45332d0c188790f90691ef112/src/hb-repacker.hh#LL72C22-L72C22
-    #[cfg(feature = "tables")]
     fn try_splitting_subtables(&mut self) {
         let splittable = self
             .objects
@@ -1096,7 +1029,6 @@ impl Graph {
         }
     }
 
-    #[cfg(feature = "tables")]
     fn split_subtables_if_needed(&mut self, lookup: ObjectId) {
         // So You Want to Split Subtables:
         // - support PairPos and MarkBase.
@@ -1223,7 +1155,6 @@ impl Graph {
     }
 }
 
-#[cfg(feature = "tables")]
 impl Default for Priority {
     fn default() -> Self {
         Priority::ZERO
