+++ write-fonts/src/util.rs
@@ -3,10 +3,8 @@
 /// Iterator that iterates over a vector of iterators simultaneously.
 ///
 /// Adapted from <https://stackoverflow.com/a/55292215>
-#[cfg(feature = "tables")]
 pub struct MultiZip<I: Iterator>(Vec<I>);
 
-#[cfg(feature = "tables")]
 impl<I: Iterator> MultiZip<I> {
     /// Create a new MultiZip from a vector of iterators
     pub fn new(vec_of_iters: Vec<I>) -> Self {
@@ -14,7 +12,6 @@ impl<I: Iterator> MultiZip<I> {
     }
 }
 
-#[cfg(feature = "tables")]
 impl<I: Iterator> Iterator for MultiZip<I> {
     type Item = Vec<I::Item>;
 
@@ -27,13 +24,11 @@ impl<I: Iterator> Iterator for MultiZip<I> {
 ///
 /// This is particularly useful when porting from Python where reading
 /// idx - 1, with -1 meaning last, is common.
-#[cfg(feature = "tables")]
 pub trait WrappingGet<T> {
     fn wrapping_next(&self, idx: usize) -> &T;
     fn wrapping_prev(&self, idx: usize) -> &T;
 }
 
-#[cfg(feature = "tables")]
 impl<T> WrappingGet<T> for &[T] {
     fn wrapping_next(&self, idx: usize) -> &T {
         &self[match idx {
@@ -58,14 +53,12 @@ impl<T> WrappingGet<T> for &[T] {
 /// - [PEP425](https://peps.python.org/pep-0485/)
 /// - [math.isclose](https://docs.python.org/3/library/math.html#math.isclose)
 #[derive(Clone, Copy, Debug)]
-#[cfg(feature = "tables")]
 pub struct FloatComparator {
     // TODO: Make it generic over T: Float? Need to add num-traits as a dependency
     rel_tol: f64,
     abs_tol: f64,
 }
 
-#[cfg(feature = "tables")]
 impl FloatComparator {
     /// Create a new FloatComparator with the specified relative and absolute tolerances.
     pub fn new(rel_tol: f64, abs_tol: f64) -> Self {
@@ -92,7 +85,6 @@ impl FloatComparator {
     }
 }
 
-#[cfg(feature = "tables")]
 impl Default for FloatComparator {
     /// Create a new FloatComparator with `rel_to=1e-9` and `abs_tol=0.0`.
     fn default() -> Self {
@@ -105,63 +97,6 @@ impl Default for FloatComparator {
 ///
 /// To use different relative or absolute tolerances, create a FloatComparator
 /// and use its `isclose` method.
-#[cfg(feature = "tables")]
 pub fn isclose(a: f64, b: f64) -> bool {
     FloatComparator::default().isclose(a, b)
-}
-
-/// Search range values used in various tables
-#[derive(Clone, Copy, Debug)]
-pub struct SearchRange {
-    pub search_range: u16,
-    pub entry_selector: u16,
-    pub range_shift: u16,
-}
-
-impl SearchRange {
-    //https://github.com/fonttools/fonttools/blob/729b3d2960ef/Lib/fontTools/ttLib/ttFont.py#L1147
-    /// calculate searchRange, entrySelector, and rangeShift
-    ///
-    /// these values are used in various places, such as [the base table directory]
-    /// and [cmap format 4].
-    ///
-    /// [the base table directory]: https://learn.microsoft.com/en-us/typography/opentype/spec/otff#table-directory
-    /// [cmap format 4]: https://learn.microsoft.com/en-us/typography/opentype/spec/cmap#format-4-segment-mapping-to-delta-values
-    pub fn compute(n_items: usize, item_size: usize) -> Self {
-        let entry_selector = (n_items as f64).log2().floor() as usize;
-        let search_range = (2.0_f64.powi(entry_selector as i32) * item_size as f64) as usize;
-        // The result doesn't really make sense with 0 tables but ... let's at least not fail
-        let range_shift = (n_items * item_size).saturating_sub(search_range);
-        SearchRange {
-            search_range: search_range.try_into().unwrap(),
-            entry_selector: entry_selector.try_into().unwrap(),
-            range_shift: range_shift.try_into().unwrap(),
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// based on example at
-    /// <https://learn.microsoft.com/en-us/typography/opentype/spec/cmap#format-4-segment-mapping-to-delta-values>
-    #[test]
-    fn simple_search_range() {
-        let SearchRange {
-            search_range,
-            entry_selector,
-            range_shift,
-        } = SearchRange::compute(39, 2);
-        assert_eq!((search_range, entry_selector, range_shift), (64, 5, 14));
-    }
-
-    #[test]
-    fn search_range_no_crashy() {
-        let foo = SearchRange::compute(0, 0);
-        assert_eq!(
-            (foo.search_range, foo.entry_selector, foo.range_shift),
-            (0, 0, 0)
-        )
-    }
 }
